+++ rest-api-server/src/main/kotlin/eu/tib/orkg/prototype/statements/domain/model/Stats.kt
@@ -20,6 +20,8 @@ data class Stats(
     val users: Long,
     val observatories: Long,
     val organizations: Long,
+    @JsonProperty("orphaned_nodes")
+    val orphanedNodes: Long,
     @JsonProperty("extras")
     val extraCounts: Map<String, Long>?
 )