+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- The number of orphaned nodes in the database is now tracked via the statistics endpoint. (Closes: [#420](https://gitlab.com/TIBHannover/orkg/orkg-backend/-/issues/420))
 
 ## [0.17.2] - 2022-11-08
 ### Fixed